コード修正

+++ app/controllers/items_controller.rb
@@ -38,6 +38,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    return unless item.user == current_user
+
     item.destroy
     redirect_to root_path
   end
